requirements.txt
streamlit
streamlit_option_menu
msal
python-dotenv
azure-identity
azure-keyvault-secrets

